Organize and make things better
