+++ .github/ISSUE_TEMPLATE/bug.yml
@@ -1,0 +1,52 @@
+name: Bug Report
+description: Report broken or incorrect behaviour
+body:
+  - type: markdown
+    attributes:
+      value: >
+        Thanks for taking the time to fill out a bug.
+        If you want real-time support, consider joining our Discord at https://discord.gg/r3sSKJJ instead.
+
+        Please note that this form is for bugs only!
+  - type: input
+    attributes:
+      label: Summary
+      description: A simple summary of your bug report
+    validations:
+      required: true
+  - type: textarea
+    attributes:
+      label: Reproduction Steps
+      description: >
+         What you did to make it happen.
+    validations:
+      required: true
+  - type: textarea
+    attributes:
+      label: Expected Results
+      description: >
+        What did you expect to happen?
+    validations:
+      required: true
+  - type: textarea
+    attributes:
+      label: Actual Results
+      description: >
+        What actually happened?
+    validations:
+      required: true
+  - type: checkboxes
+    attributes:
+      label: Checklist
+      description: >
+        Let's make sure you've properly done due dilligence when reporting this issue!
+      options:
+        - label: This bug was found on the latest version of EpikCord
+          required: true
+        - label: You have checked that there is no issue already mentioning this bug, open or closed
+          required: true
+        - label: Full details of the bug have been provided/shown, including full tracebacks in case of exceptions
+  - type: textarea
+    attributes:
+      label: Additional Context
+      description: If there is anything else to say, please do so here.
