+++ .github/ISSUE_TEMPLATE/config.yml
@@ -1,0 +1,5 @@
+blank_issues_enabled: false
+contact_links:
+  - name: Development Discord Server
+    about: Use our development Discord server to ask for questions.
+    url: https://discord.gg/4R473R73kQ
