+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,19 @@
+## Description
+
+<!--Describe your pull request here-->
+
+## Checklist
+
+<!--To check an item in the checklist, put an x between the square brackets without spaces. Like - [x]-->
+
+- [ ] - All code changes have been tested, if any
+- [ ] - This PR fixes an already existing issue
+- [ ] - This PR adds a new feature (new functions, methods, classes, parameters, etc.)
+- [ ] - This PR fixes a bug
+- [ ] - This PR is a breaking change (removes/renames functions, parameters etc.)
+- [ ] - This PR is NOT a code change (changes Documentation files, README.md, CONTRIBUTING.md etc.)
+- [ ] - This PR is a code refactor
+
+## Additional context
+
+<!--This is optional, provide some more context on your pull request here-->
