+++ .github/workflows/bandit.yml
@@ -1,0 +1,10 @@
+name: bandit
+on: [pull_request, push]
+jobs:
+  bandit:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - uses: actions/setup-python@v2
+      - run: pip install bandit
+      - run: bandit --recursive .
