+++ bot/__main__.py
@@ -1,0 +1,6 @@
+from .utils import RoboEpik, Config
+from EpikCord import Intents
+
+
+if __name__ == "__main__":
+    RoboEpik(Config.token, Intents().all).login()
