+++ bot/utils/__init__.py
@@ -1,0 +1,2 @@
+from .bot import *
+from .config import *
