+++ readme.md
@@ -1,0 +1,40 @@
+# RoboEpik
+
+Welcome to the **official** repository of RoboEpik.
+
+
+## Contents
+
+- [RoboEpik](#RoboEpik)
+  - [Introduction](#introduction)
+  - [Bug/Typo/Mistake](#bugtypomistake)
+  - [Contributing](#contributing)
+
+
+### Introduction
+
+RoboEpik is a multi purpose bot for EpikCord.py RoboEpik is the main (first) bot of [Epikcord.py](). It is created to serve a testing bot in the [Epikcord.py Discord Server]().
+
+Our main RoboEpik server is the Epikcord.py server. 
+
+
+<a href="https://discord.gg/4R473R73kQ" target="_blank" rel="noopener noreferrer">
+    <img src="https://img.shields.io/discord/937364424208039957?color=blue&label=Epikcord.py&logo=discord&style=for-the-badge" alt="Discord server">
+</a>
+
+
+
+### Bug/Typo/Mistake
+
+You have found a mistake or typo or bug! You can report it in our issue sections. If you know how to fix it consider following the steps in [Contributing](#contributing). 
+
+
+### Contributing
+
+**This project is only a draft. Until we get a stable Epikcord.py version then bot project will resume.**
+
+1. Create a fork of the repo
+2. Edit and make changes
+3. Open a pull request 
+
+If you don't understand, read more [here]()
