+++ requirements.txt
@@ -1,0 +1,1 @@
+git+https://github.com/EpikCord/EpikCord.py
